Moved headers from src to interface in FWCore/Framework
These headers were being used by other packages.
