+++ FWCore/Framework/interface/SharedResourcesRegistry.h
