+++ FWCore/Framework/src/ModuleFactory.cc
@@ -15,7 +15,7 @@
 // user include files
 #include "FWCore/Framework/interface/ModuleFactory.h"
 #include "FWCore/Framework/interface/EventSetupProvider.h"
-#include "FWCore/Framework/src/EventSetupsController.h"
+#include "FWCore/Framework/interface/EventSetupsController.h"
 
 //
 // constants, enums and typedefs
