+++ FWCore/Framework/src/SourceFactory.cc
@@ -14,7 +14,7 @@
 
 // user include files
 #include "FWCore/Framework/interface/SourceFactory.h"
-#include "FWCore/Framework/src/EventSetupsController.h"
+#include "FWCore/Framework/interface/EventSetupsController.h"
 #include "FWCore/Utilities/interface/EDMException.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
