+++ FWCore/TestProcessor/src/TestProcessor.cc
@@ -26,11 +26,11 @@
 #include "FWCore/Framework/interface/PathsAndConsumesOfModules.h"
 #include "FWCore/Framework/interface/RunPrincipal.h"
 #include "FWCore/Framework/interface/ESRecordsToProxyIndices.h"
-#include "FWCore/Framework/src/EventSetupsController.h"
-#include "FWCore/Framework/src/globalTransitionAsync.h"
-#include "FWCore/Framework/src/streamTransitionAsync.h"
+#include "FWCore/Framework/interface/EventSetupsController.h"
+#include "FWCore/Framework/interface/globalTransitionAsync.h"
+#include "FWCore/Framework/interface/streamTransitionAsync.h"
 #include "FWCore/Framework/interface/TransitionInfoTypes.h"
-#include "FWCore/Framework/src/ProductPutterBase.h"
+#include "FWCore/Framework/interface/ProductPutterBase.h"
 #include "FWCore/Framework/interface/DelayedReader.h"
 
 #include "FWCore/ServiceRegistry/interface/ServiceRegistry.h"
